Add signup logic
